wlan_qcm6490: Update default driver as ath11k
Current wlan script behavior is to load the qcacld wlan
driver by default.

Also, for qcacld driver, this script will do insmod and rmmod,
while the upstream driver shall be loaded as part of kernel.

This Change is to make MODNAME=ath11k_ahb by default and
not to start the dhcpcd explicitly since it is already
started by dhcpcd systemd service

Also, does not do rmmod for ath11k driver as this script
has not loaded it.

Change-Id: I954387129750054c6ef834e6f43b4415dd6633d4
CRs-Fixed: 3924182